+++ apps/next-app/src/app/dashboard/layout.tsx
@@ -1,0 +1,13 @@
+import NavBar from "@/components/Common/NavBar";
+import { SignedIn } from "@clerk/nextjs";
+
+export default function DashboardLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+        <SignedIn>
+            <NavBar />
+            <main>{children}</main>
+        </SignedIn>
+    </>
+  );
+}
