+++ apps/next-app/src/app/layout.tsx
@@ -1,14 +1,9 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import NavBar from "@/components/Common/NavBar";
 import {
   ClerkProvider,
-  SignedIn,
-  SignedOut,
 } from '@clerk/nextjs'
-import Hero from "@/components/Landing/Hero";
-import Footer from "@/components/Landing/Footer";
 import {dark} from "@clerk/themes"
 
 const geistSans = Geist({
@@ -35,13 +30,6 @@ export default function RootLayout({
     <ClerkProvider
       appearance={{
         baseTheme: dark,
-        // variables: {
-        //   colorPrimary: "#a21caf", // purple
-        //   colorBackground: "#18181b", // dark background
-        //   colorText: "#fff", // white text
-        //   colorInputBackground: "#27272a",
-        //   colorInputText: "#fff",
-        // },
         elements: {
           card: "rounded-2xl shadow-lg",
         },
@@ -51,14 +39,7 @@ export default function RootLayout({
         <body
           className={`${geistSans.variable} ${geistMono.variable} antialiased`}
         >
-          <SignedOut>
-            <Hero/>
-            <Footer/>
-          </SignedOut>
-          <SignedIn>
-            <NavBar />
-            {children}
-          </SignedIn>
+          {children}
         </body>
       </html>
     </ClerkProvider>
